+++ tools/build_cli_debug.ps1
@@ -1,5 +1,5 @@
-dotnet clean "../src/CommandLine.sln" ` -c Debug
+dotnet clean "$PSScriptRoot/../src/CommandLine.sln" ` -c Debug
 
-dotnet build "../src/CommandLine.sln" -c Debug -v minimal /p:RoslynatorDotNetCli=true /m
+dotnet build "$PSScriptRoot/../src/CommandLine.sln" -c Debug -v minimal /p:RoslynatorDotNetCli=true /m
 
 Write-Host "DONE"